+++ tests/conftest.py
@@ -98,6 +98,29 @@ def data_instance():
 
 
 @pytest.fixture
+def dummy_carbon_data():
+    """Creates a dummy carbon data object for use in tests."""
+
+    from virtual_rainforest.core.data import Data
+    from virtual_rainforest.core.grid import Grid
+
+    # Setup the data object with four cells.
+    grid = Grid(cell_nx=4, cell_ny=1)
+    data = Data(grid)
+
+    # Add the required data.
+    data["maom"] = DataArray([2.5, 1.7, 4.5, 0.5], dims=["cell_id"])
+    data["lmwc"] = DataArray([0.05, 0.02, 0.1, 0.005], dims=["cell_id"])
+    data["pH"] = DataArray([3.0, 7.5, 9.0, 5.7], dims=["cell_id"])
+    data["bulk_density"] = DataArray([1350.0, 1800.0, 1000.0, 1500.0], dims=["cell_id"])
+    data["soil_moisture"] = DataArray([0.5, 0.7, 0.6, 0.2], dims=["cell_id"])
+    data["soil_temperature"] = DataArray([35.0, 37.5, 40.0, 25.0], dims=["cell_id"])
+    data["percent_clay"] = DataArray([80.0, 30.0, 10.0, 90.0], dims=["cell_id"])
+
+    return data
+
+
+@pytest.fixture
 def new_axis_validators():
     """Create new axis validators to test methods and registration."""
     from virtual_rainforest.core.axes import AxisValidator
