+++ tests/test_main.py
@@ -12,13 +12,7 @@ import pytest
 from numpy import datetime64, timedelta64
 
 from virtual_rainforest.core.base_model import BaseModel, InitialisationError
-from virtual_rainforest.main import (
-    check_for_fast_models,
-    configure_models,
-    extract_timing_details,
-    select_models,
-    vr_run,
-)
+from virtual_rainforest.main import vr_run
 from virtual_rainforest.models.soil.soil_model import SoilModel
 
 from .conftest import log_check
@@ -73,6 +67,7 @@ from .conftest import log_check
 )
 def test_select_models(caplog, model_list, no_models, raises, expected_log_entries):
     """Test the model selecting function."""
+    from virtual_rainforest.main import select_models
 
     with raises:
         models = select_models(model_list)
@@ -130,14 +125,15 @@ def test_select_models(caplog, model_list, no_models, raises, expected_log_entri
     ],
 )
 def test_configure_models(
-    caplog, data_instance, config, output, raises, expected_log_entries
+    caplog, dummy_carbon_data, config, output, raises, expected_log_entries
 ):
     """Test the function that configures the models."""
+    from virtual_rainforest.main import configure_models, select_models
 
     with raises:
         model_list = select_models(["soil"])
 
-        models = configure_models(config, data_instance, model_list)
+        models = configure_models(config, dummy_carbon_data, model_list)
 
         if output is None:
             assert models == [None]
@@ -313,6 +309,7 @@ def test_vr_run_bad_model(mocker, caplog):
 )
 def test_extract_timing_details(caplog, config, output, raises, expected_log_entries):
     """Test that function to extract main loop timing works as intended."""
+    from virtual_rainforest.main import extract_timing_details
 
     with raises:
         current_time, update_interval, end_time = extract_timing_details(config)
@@ -342,6 +339,7 @@ def test_extract_timing_details(caplog, config, output, raises, expected_log_ent
 )
 def test_check_for_fast_models(caplog, update_interval, expected_log_entries):
     """Test that function to warn user about short module time steps works."""
+    from virtual_rainforest.main import check_for_fast_models
 
     # Create SoilModel instance and then populate the update_interval
     model = SoilModel.__new__(SoilModel)
